feat: bind overlap events for item activation

+++ Source/ThisIsSparta/Public/HealingItem.h
@@ -13,7 +13,7 @@ class THISISSPARTA_API AHealingItem : public ABaseItem
 	AHealingItem();
 	
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = Item)
-	float HealAmount;
+	int32 HealAmount;
 	
 	virtual void ActivateItem(AActor* Activator) override;
 };
